rename: type 파일 위치 수정 (components 폴더 > src 폴더)
